Refactoring classnames in sidebar First preparations for the PR
