+++ packages/vizbuilder/src/components/ConditionManager.css
@@ -1,0 +1,64 @@
+
+.condition-item {
+  padding: 0.3rem 0.3rem 0.1rem;
+  margin: 0 0 0.4rem;
+  border: 1px solid;
+
+  &.disabled {
+    opacity: 0.7;
+  }
+
+  & > .group {
+    margin: 0 0 0.2rem;
+  }
+
+  & .label {
+    display: block;
+    text-align: center;
+    text-transform: uppercase;
+    letter-spacing: 1px;
+  }
+
+  &.closed {
+    display: flex;
+    flex-direction: row;
+
+    & .condition-content {
+      flex: 1 0;
+      list-style: none;
+      margin: 0;
+      padding: 0;
+
+      & li {
+        margin-bottom: 0.2rem;
+      }
+    }
+
+    & .condition-oper {
+      text-transform: uppercase;
+      font-size: 0.8em;
+    }
+
+    & .condition-value span {
+      display: inline-block;
+      margin-right: 3px;
+      padding: 0 2px;
+      font-size: 0.9em;
+      text-transform: uppercase;
+      border: 1px solid rgba(0, 0, 0, 0.1);
+    }
+
+    & .condition-action {
+      flex: 0 0;
+      align-self: flex-start;
+    }
+  }
+}
+
+.condition-actions {
+  text-align: right;
+
+  & .pt-button {
+    margin-left: 0.2rem;
+  }
+}
