+++ packages/vizbuilder/src/index.css
@@ -4,28 +4,20 @@
   flex-direction: row;
 
   & > .area-loading {
-    position: fixed;
-    top: 0;
-    left: 0;
-    z-index: 100;
-    width: 100%;
-    max-width: none;
-    height: 100vh;
-    opacity: 0;
-    visibility: hidden;
+    display: none;
   }
 
   &.loading > .area-loading {
-    opacity: 1;
-    visibility: visible;
+    display: flex;
   }
 
   & > .area-sidebar {
-    flex: 0 0 300px;
+    flex: 0 0 auto;
+    width: 300px;
   }
 
   & > .area-chart {
-    flex: 1 0;
+    flex: 1 0 0;
   }
 }
 
@@ -36,5 +28,31 @@
 
   & > .wrapper {
     margin: 15px;
+  }
+}
+
+.area-loading {
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: 200;
+  width: 100%;
+  height: 100%;
+  max-width: none;
+
+  color: var(--white);
+  background-color: rgba(0, 0, 0, 0.75);
+
+  & .pt-non-ideal-state-visual {
+    width: 100%;
+    max-width: 200px;
+  }
+
+  & .pt-progress-meter {
+    background-color: var(--light);
+  }
+
+  & .pt-non-ideal-state-title {
+    color: var(--white);
   }
 }
